requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
blinker==1.7.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.15.0
httpx==0.23.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.2
openai==1.3.4
packaging==23.2
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
scikit-learn==1.3.2
scipy==1.11.4
sklearn==0.0.post11
sniffio==1.3.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
